Lab1B
This one was a little tricker. I saw that the test function was called with arg - 0x1337D00D. Then there was a switch statement to the decrypt function which ended up being a simple and of each character in an encrypted string with the switch value. The goal was to make the encrypted string say "Congratulations!" which ended up needing a value of 0x12 or decimal 18.