Lab1A
I realized that in order to be authenticated, the serial number had to match the username after a set of operations had been performed. The user name had to be longer than 5 characters, and could not include any characters < 31. The set of operations was tricky, but basically set the last byte of the username to 0's, took the 4th char of the username ^ 0x1337 + 0x5EEDED and used this for the start of the sum. Then it would go through the username string char by char and and add the sum ^ curChar and mod it by 1337.

Eventually I managed to suss out the pattern by keeping the username consistent ('ddddddd') and working through the program by hand.