Lab2B
This one was trickier. I messed around in GDB until I found out I needed exactly 27 bytes of overflow until I could overwrite the return pointer to point to the shell function. Once I did that, i still needed to figure out how to setup the stack with the proper argument for the system call. I realized that the format of functions allowed me to overwrite the shell functions return pointer with junk data and then specify the argument, which in this case was the address of the string "/bin/sh".